pyproject.toml
[project]
name = "2026se-task1-thribhu-k"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "Flask",
    "bcrypt",
    "flask_wtf",
    "flask-csp",
    "requests",
    "flask-jwt-extended>=4.7.1",
    "python-dotenv>=1.2.1",
    "flask-cors>=6.0.1",
]
